requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --no-annotate
#
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
commonmark==0.9.1
cython==0.29.28
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
jinja2==3.1.1
markupsafe==2.1.1
marshmallow==3.17.0
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
prison==0.2.1
prompt-toolkit==3.0.29
pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-graphql-client==0.4.3
pytz==2022.1
pyyaml==6.0
requests==2.27.1
rich==12.3.0
six==1.16.0
soupsieve==2.3.2.post1
sqlalchemy==1.4.35
sqlparse==0.4.2
tabulate==0.8.9
typing-extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
yarl==1.7.2
